Add feature to display user profile picture
Added a new API endpoint to retrieve a user's profile picture by their ID. The endpoint returns the URL of the user's profile picture if it exists, otherwise it returns a default image. This feature was added in response to customer feedback requesting an easier way to access user profile pictures.